Fix compilation when CMAKE_C_FLAGS is set
CMake would insert a semicolon between the contents of CMAKE_C_FLAGS, and the newly-appended flags. This goes unnoticed if CMAKE_C_FLAGS isn't set to anything, but when it is, it causes the compiler to throw a fit. I encountered this while trying to make an Arch Linux package (which I wasn't able to, since the CMake script lacks install targets).