Temporary fix for SDL player bug.
Commit e0d4316111b0c9a04f0e22f640a55d1ff4821c49 introduced a bug
where the SDL MIDI player would be silent. A temporary fix has
been applied onto that bug, where the commit's fix is skipped if
the MIDI player selected is currently the SDL one.

Please note that this is only supposed to be a temporary fix, and
a placeholder for a more definitive one. Also, other MIDI players
have not been tested yet.