Reverted more slight formatting-related changes
Those few had also slipped by.

Please note for posterity that functions copied to new files
(in gus-experimental branch, which is not PR-ready yet) due
to being static will not have the same formatting as the
original ones. The original ones will be left untouched, though.

This is not relevant to the gus-dev branch (more closely PRBoom-specific,
with a focus on fixing bugs and helping implementing actually useful
features, rather than experimental ones).