Update INSTALL
I have changed the "Basic Installation" to the top as the SVN isn't the main place I believe people get the code from, and changed the Basic Install guide to have the correct instructions (with ./configure missing and instead being cmake .) and added some basic instructions for manual installation (copying the files instead of using make install) as well as a link to FreeDoom as a potential alternative if the user doesn't have iD's Doom. 

The required libraries are correct on Ubuntu and Debian systems (tested on Ubuntu 20.04) and with minor changes to packages downloaded, these instructions work on Void Linux too.