get rid of the "Screen Multiple Factor" feature
Since the switch to SDL2 we have seamless scaling of the game window.
This means, we don't need to explicitly multiply pixels in order to
scale a low resolution up to the desired window size anymore. This
works for all rendering modes, both windowed and fullscreen - except
for OpenGL rendering, but this doesn't support the "screen multiple
factor" feature either.