add a compatibility option for MBF sky transfers
This is a cosmetic change and considered safe for all complevels, e.g.
Crispy Doom applies it without even asking. Thus, in order to keep
things simple and backward compatible, we don't go the long way of
adding this to the comp[] and default_comp[] arrays, but add a simple
boolean variable. Also, default this to `true` for the aforementioned
reasons, purists may still disable it if deemed necessary.