Remove ALSA sched-queue timer resolution code
The music stutering issues were found to be unrelated to the resolution
of the timer in ALSA's scheduling queues. This was determined by setting
the resolution to 5 whopping milliseconds for a single test run, which did
not affect the quality of playback in any manner, thus ruling out timer
resolution as a possible cause.

Investigation is still underway to pinpoint a better solution for this
music stutter issue.