allow colored blood to get set/overridden by DEHACKED
This introduces the .blood field in mobjinfo[] which points to a
(virtual) blood mapthing type. By default, point this to MT_BLUEBLOOD
for Cacodemons and MT_GREENBLOOD for Hell nobles and MT_BLOOD for
anything else.

While at it, also introduce the MT_FUZZYBLOOD type for spectre blood,
but do not set it for spectres by default (yet?).

Two questions left:
- Should we initialize to MT_BLOOD for all other monsters or rather
  MT_NULL? This would leave the possibility to apply a special third
  colorization with a crafted RED color translation table.
- Should we unapply the blood assignment if we detect that a monster
  has been dehacked?