Do not cast degenmobj_t as mobj_t, breaks some aliasing rules.
This is an issue on some platforms where gcc can assume that mobj_t
pointers are 64 bit aligned (due to `flags` being a 64 bit int) and
degenmobj_t pointers being only 32 bit aligned.
In said platforms gcc will do optimizations that result in mis-aligned
memory loads, which may or may not work.

Can provide more info on this, but it breaks builds on MIPS64 CPUs.