Sound Quirks Fix
Changes an `or` statement to an `and` in a check for stopping sound. The 'kill old sound' code is supposed to emulate a quirk with item pickup sounds however it was applying the quirk to all sounds. Changing the 'or' to an 'and' fixes this. I know that the sound quirks was changed to be a setting separated from -complevel but I think it'd still be a nice addition to have this fix for the quirks.