Fullbright fixes for v1.2 compat - take 2
In versions before v1.4 the Spiderdemon didn't go fullbright while attacking. Note that this means that the whole sprite didn't light up, not to be confused with the (faint) muzzle flashes around its chainguns which were there since day one.

The Soulsphere went fullbright in v1.25 (the Sybex-exclusive shareware). Starting with v1.4 beta all powerups are rendered fullbright.

Credit, as usual, to SmileTheory for the original Choco implementation.