Don't overwrite user-configured custom docdir (#480)
We want to provide a default docdir for PrBoom-Plus, but let's not
overwrite the custom docdir a user may provide.

We can't simply set the ${CMAKE_INSTALL_DOCDIR} in the CMakeLists.txt
file because the GNUInstallDirs has already set the
${CMAKE_INSTALL_DOCDIR} cache variable before a later set command; that
later set command is prevented from changing an already set cache
variable. The way to solve this is to check if ${CMAKE_INSTALL_DOCDIR}
is set before the include(GNUInstallDirs), and then only set the default
location when needed.